Corrected computation of JHOUR in getdte - if a model started at hour 23.5 the old formula led to a segmentation fault, otherwise only the diurnal diagnostics were affected by that error, if the model started at a half hour. (Runs in which NSSW was set to an even number were not affected).

+++ model/MODEL_COM.f
@@ -709,7 +709,7 @@ C**** keep track of min/max time over the combined diagnostic period
         Jmn=Jmn+1
       end do
       Jdate=Jd-JDendOfM(Jmn-1)
-      Jhour=nint(mod(It*hrday/Nday,hrday))
+      Jhour=mod(It*hrday/Nday,hrday)
       amn=amonth(Jmn)
 
       return
